--- src/headers/Classes/popup_message_window.hpp
@@ -1,18 +1,0 @@
-class popup_window {
-public:
-
-popup_window() {}
-
-void setAttributes(int size_x, int size_y, std::string message_text)
-{
-
-}
-
-private:
-
-sf::RectangleShape background;
-
-sf::Text message_text;
-
-
-};
