--- src/headers/Functions/Settings/defaults/dafault_paths.hpp
@@ -1,6 +1,0 @@
-void setDefault_paths()
-{
-    settings_defaults::steam_gamedir_def();
-    settings_defaults::instances_dir_def();
-    settings_defaults::steamcmd_dir_def();
-}
